% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_analyse_MPI.R
\name{analyse_mpi}
\alias{analyse_mpi}
\title{Produz os graficos e tabelas relacionados aos resultados do MPI}
\usage{
analyse_mpi(
  dt = NULL,
  grupos = "all",
  k_ref = 0.33,
  cutoffs = c(0.1, 0.2, 0.25, 0.33, 0.4, 0.5),
  anos_tabs = c(1981, 1987, 1993, 1999, 2005, 2011, 2017, 2025),
  anos_comp = c(1985, 1995, 2005, 2015, 2025),
  anos_maps = c(1981, 1995, 2009, 2025),
  anos_dens = c(1985, 1995, 2005, 2015, 2025),
  anos_beta = c(1981, 2025),
  anos_decp = list(c(1981, 1986), c(1986, 1995), c(1995, 2002), c(2002, 2008), c(2008,
    2014), c(2014, 2019), c(2019, 2025)),
  dpi = 200,
  dir_out = "output/graphs",
  shapefiles = NULL,
  base_size = 10,
  usar_fonte = T,
  ext = "png",
  verbose = T
)
}
\arguments{
\item{dt}{Banco de dados que foi carregado, pelo usuario, ao ler o arquivo \code{pnad_completa.parquet}}

\item{grupos}{Tipos de graficos que serao produzidos pela funcao}

\item{k_ref}{Valor escolhido (default = 0.33) para apresentacao do report}

\item{cutoffs}{Valores de cutoff K utilizados para os calculos do binario de pobreza}

\item{anos_tabs}{Anos escolhidos para apresentar nas tabelas (default = de 1981 em diante, a cada 6 anos, alem do ano mais recente)}

\item{anos_comp}{Anos escolhidos para comparativos (default = todo ano terminado em 5, alem do ano mais recente)}

\item{anos_maps}{Anos escolhidos para criacao dos mapas (default = de 1981 em diante, a cada 14 anos, alem do ano mais recente)}

\item{anos_dens}{Anos escolhidos para analises de dominancia e densidade (default = todo ano terminado em 5, alem do ano mais recente)}

\item{anos_decp}{Lista de pares de ano para analise das decomposicoes (default = 1981-1990, 1990-2001, 2001-2011, 2011-2019, 2019-2025)}

\item{dir_out}{Local onde os objetos serao salvos (default = 'output')}
}
\value{
data.table unico, com calculos de pobreza unidimensional (RPC) e multidimensional (MPI usando cutoff K)
}
\description{
Produz os graficos e tabelas relacionados aos resultados do MPI
}
