% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_create_MPI.R
\name{create_mpi}
\alias{create_mpi}
\title{Calcula o MPI-LA para todos os anos da PNAD (anual ou continua) disponiveis}
\usage{
create_mpi(
  input_dir = "data/03_filtered",
  output_file = "data/pnad_completa.parquet",
  ks = c(0.1, 0.2, 0.25, 0.33, 0.4, 0.5),
  anos = NULL
)
}
\arguments{
\item{input_dir}{Caminho logico onde se encontram os arquivos anuais ja filtrados, que serao tratados pelas funcoes \code{rowcalc} e \code{aggcalc}}

\item{output_file}{Nome e local do diretorio onde sera salva a base empilhada (com nome padrao \code{pnad_completa.parquet})}

\item{ks}{Valores de cutoff K utilizados para os calculos do binario de pobreza}

\item{anos}{Nulo por default, permite criar a base so com anos selecionados}
}
\value{
data.table unico, com todos os anos empilhados, e todos os indicadores e cutoffs calculados
}
\description{
Calcula o MPI-LA para todos os anos da PNAD (anual ou continua) disponiveis
}
