% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_dashboard_MPI.R
\name{dashboard_mpi}
\alias{dashboard_mpi}
\title{Cria o dashboard de análise do MPI-Brasil}
\usage{
dashboard_mpi(
  path_dict = "data/mpi_dictionary.rds",
  path_pnad = "data/pnad_completa.parquet",
  path_models = "output/models",
  port = NULL,
  launch = TRUE
)
}
\arguments{
\item{path_dict}{Caminho onde o dicionário está salvo (defalt = 'data/mpi_dictionary.rds')}

\item{path_pnad}{Caminho onde a base PNAD está salva (defalt = 'data/pnad_completa.parquet')}

\item{path_models}{Caminho onde estão salvos os modelos estatísticos (default = 'output/models')}

\item{port}{Porta utilizada para a apresentação do dashboard, no painel}

\item{launch}{Booleano (default = T) indicando se o dashboard será visualizado}
}
\description{
Cria o dashboard de análise do MPI-Brasil
}
