% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_resume_MPI.R
\name{resume_mpi}
\alias{resume_mpi}
\title{Cria tabela-resumo dos resultados do MPI-LA, conforme parametros indicados}
\usage{
resume_mpi(
  dt = NULL,
  grupos = "ano",
  k_output = 0.33,
  ks = c(0.1, 0.2, 0.25, 0.33, 0.4, 0.5)
)
}
\arguments{
\item{dt}{Banco de dados que foi carregado, pelo usuario, ao ler o arquivo \code{pnad_completa.parquet}}

\item{grupos}{Nome dos atributos (default = 'ano') pelos quais serao feitas as analises segmentadas}

\item{k_output}{Valor escolhido (default = 0.33) para apresentacao do report}

\item{ks}{Valores de cutoff K utilizados para os calculos do binario de pobreza}
}
\value{
data.table unico, com calculos de pobreza unidimensional (RPC) e multidimensional (MPI usando cutoff K)
}
\description{
Cria tabela-resumo dos resultados do MPI-LA, conforme parametros indicados
}
