% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_cluster_MPI.R
\name{cluster_mpi}
\alias{cluster_mpi}
\title{Realiza a clusterização hierárquica (Ward) sobre os 5 scores dimensionais do MPI-LA}
\usage{
cluster_mpi(dt, k, reduce = TRUE, dicts = NULL)
}
\arguments{
\item{dt}{Tabela agregada produzida pelo pipeline prep_cluster, já filtrada pelo usuário}

\item{k}{Número de clusters (inteiro, obrigatório)}
}
\value{
Lista de objetos relacionados à clusterização
}
\description{
Realiza a clusterização hierárquica (Ward) sobre os 5 scores dimensionais do MPI-LA
}
