% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_download_PNAD.R
\name{download_pnad}
\alias{download_pnad}
\title{Baixa arquivos compactados da PNAD Anual a partir do CEM/USP}
\usage{
download_pnad(anos, destino = "data/01_raw/pnad")
}
\arguments{
\item{anos}{Vetor de anos a baixar (ex: c(1981, 1990, 2001))}

\item{destino}{Pasta de destino (default: 'data/01_raw/pnad')}
}
\description{
Baixa arquivos compactados da PNAD Anual a partir do CEM/USP
}
